[SCRUM-60] rename: auth 디렉토리 global -> domain 이동
