+++ src/main/java/com/projectlyrics/server/domain/auth/external/kakao/dto/KakaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.projectlyrics.server.global.auth.external.kakao.dto;
+package com.projectlyrics.server.domain.auth.external.kakao.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
