+++ src/main/java/com/projectlyrics/server/domain/user/repository/QueryDslQueryUserRepository.java
@@ -2,7 +2,7 @@ package com.projectlyrics.server.domain.user.repository;
 
 import com.projectlyrics.server.domain.user.entity.QUser;
 import com.projectlyrics.server.domain.user.entity.User;
-import com.projectlyrics.server.global.auth.external.AuthProvider;
+import com.projectlyrics.server.domain.auth.external.AuthProvider;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
