+++ src/main/java/com/projectlyrics/server/domain/user/service/UserCommandService.java
@@ -1,7 +1,7 @@
 package com.projectlyrics.server.domain.user.service;
 
 import com.projectlyrics.server.domain.user.dto.response.LoginResponse;
-import com.projectlyrics.server.global.auth.external.dto.request.UserLoginRequest;
+import com.projectlyrics.server.domain.auth.external.dto.request.UserLoginRequest;
 
 public interface UserCommandService {
