+++ src/main/java/com/projectlyrics/server/domain/user/service/impl/UserCommandServiceImpl.java
@@ -4,12 +4,12 @@ import com.projectlyrics.server.domain.user.dto.response.LoginResponse;
 import com.projectlyrics.server.domain.user.entity.User;
 import com.projectlyrics.server.domain.user.repository.CommandQueryUserRepository;
 import com.projectlyrics.server.domain.user.service.UserCommandService;
-import com.projectlyrics.server.global.auth.authentication.UserAuthentication;
-import com.projectlyrics.server.global.auth.external.dto.request.UserLoginRequest;
-import com.projectlyrics.server.global.auth.external.dto.response.UserInfoResponse;
-import com.projectlyrics.server.global.auth.external.kakao.KakaoSocialService;
-import com.projectlyrics.server.global.auth.jwt.JwtTokenProvider;
-import com.projectlyrics.server.global.auth.jwt.redis.service.TokenService;
+import com.projectlyrics.server.domain.auth.authentication.UserAuthentication;
+import com.projectlyrics.server.domain.auth.external.dto.request.UserLoginRequest;
+import com.projectlyrics.server.domain.auth.external.dto.response.UserInfoResponse;
+import com.projectlyrics.server.domain.auth.external.kakao.KakaoSocialService;
+import com.projectlyrics.server.domain.auth.jwt.JwtTokenProvider;
+import com.projectlyrics.server.domain.auth.jwt.redis.service.TokenService;
 import com.projectlyrics.server.global.error_code.ErrorCode;
 import com.projectlyrics.server.global.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
