--- src/main/java/com/projectlyrics/server/global/auth/external/service/SocialService.java
@@ -1,9 +1,0 @@
-package com.projectlyrics.server.global.auth.external.service;
-
-import com.projectlyrics.server.global.auth.external.dto.request.UserLoginRequest;
-import com.projectlyrics.server.global.auth.external.dto.response.UserInfoResponse;
-
-public abstract class SocialService {
-
-  public abstract UserInfoResponse login(UserLoginRequest loginRequest);
-}
